move bytecodes classes to separatre file
